Add a simple serialization mechanism.
This will mainly be for testing, specifically to transfer data to
and from the GPU.

The `Serializer<T>` class implements a binary serialization that
can write to (`serialize`) and read from (`deserialize`) a byte
buffer.  Also added convenience routines for serializing
a list of parameters.

This will mainly be for testing, specifically to transfer data to
and from the GPU.