Set nonzeroPivot threshold to zero in ColPivHouseholderQR.
This is more consistent with the documentation for `nonzeroPivots()`,
which already explicitly states we main nonzero in the "exact" sense,
not a fuzzy sense. This is also more consistent with `FullPivLU`, which
also compares pivots to exactly zero. Transitioned to using `rank()` in
solves that *mostly* restores the original behaviour, is also more
consistent with `FullPivLU`, but allows the user-defined threshold via
`setThreshold()` to be used in controlling pivots to consider in `solve()`.

Otherwise, the issue is that this hard-coded threshold is not
configurable, which can lead to poor results on some poorly conditioned
matrices (e.g. #2318).

Fixes #2318.